requirements.txt
aiohttp~=3.8.4
aiosignal~=1.3.1
async-timeout~=4.0.2
attrs~=23.1.0
cachetools~=5.3.1
certifi~=2023.5.7
charset-normalizer~=3.1.0
dataclasses-json~=0.5.7
frozenlist~=1.3.3
google-api-core~=2.11.0
google-api-python-client~=2.87.0
google-auth~=2.19.0
google-auth-httplib2~=0.1.0
google-search-results~=2.4.2
googleapis-common-protos~=1.59.0
httplib2~=0.22.0
idna~=3.4
langchain~=0.0.181
marshmallow~=3.19.0
marshmallow-enum~=1.5.1
multidict~=6.0.4
mypy-extensions~=1.0.0
numexpr~=2.8.4
numpy~=1.24.3
openai~=0.27.7
openapi-schema-pydantic~=1.2.4
packaging~=23.1
protobuf~=4.23.2
pyasn1~=0.5.0
pyasn1-modules~=0.3.0
pydantic~=1.10.8
pyparsing~=3.0.9
python-dotenv~=1.0.0
pytube~=15.0.0
PyYAML~=6.0
requests~=2.31.0
rsa~=4.9
six~=1.16.0
SQLAlchemy~=2.0.15
tenacity~=8.2.2
tqdm~=4.65.0
typing-inspect~=0.9.0
typing_extensions~=4.6.2
uritemplate~=4.1.1
urllib3~=1.26.15
yarl~=1.9.2
youtube-transcript-api~=0.6.0
